refactor: add mixin folder and string extension mixin into that
